+++ package.json
@@ -25,10 +25,10 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.2",
     "@testing-library/react": "^10.4.8",
-    "@testing-library/user-event": "^12.1.0",
+    "@testing-library/user-event": "^12.1.1",
     "@types/jest": "^26.0.9",
     "@types/node": "^14.0.27",
-    "@types/react": "^16.9.44",
+    "@types/react": "^16.9.45",
     "@types/react-dom": "^16.9.8",
     "@types/react-redux": "^7.1.9",
     "cross-env": "^7.0.2",
