+++ src/App.tsx
@@ -5,15 +5,15 @@ import { useDispatch, useSelector } from "react-redux"
 import { ToastContainer, toast, Slide } from "react-toastify"
 import { ReactFlowProvider } from "react-flow-renderer"
 import { toggleEditMode, toggleDelSelected, selectSideLeft } from "./features/ux/uxSlice"
-import AudioGraph from "./components/graph/AudioGraph"
-import MenuOpener from "./components/side/MenuOpener"
-import Analysers2D from "./components/side/Analysers2D"
+import AudioGraph from "./components/AudioGraph"
+import MenuOpener from "./components/Menu/MenuOpener"
+import Analysers2D from "./components/Analyser2D"
 import { Main, SideBar, WidgetRows } from "./styled"
 import Sequencer from "./components/Sequencer"
-import Piano from "./components/side/Piano"
-import Brand from "./components/side/Brand"
-import Load from "./components/side/Load"
-import Save from "./components/side/Save"
+import Piano from "./components/Piano"
+import Brand from "./components/misc/Brand"
+import Load from "./components/LoadSave/Load"
+import Save from "./components/LoadSave/Save"
 
 const keyMap = {
   TOGGLE_EDIT_MODE: "m",
