+++ src/components/Analyser2D/Analysers2D.tsx
