+++ src/components/Analyser2D/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Analysers2D"
