+++ src/components/AudioGraph/elems/AudioParamsView.tsx
@@ -2,7 +2,7 @@
 import { jsx } from "@emotion/core"
 import { useEffect, useState, Fragment } from "react"
 import { AudioParamSetting } from "./AudioParamForm"
-import { DataRow, DataKey, DataNote, H2 } from "../elems/styled"
+import { DataRow, DataKey, DataNote, H2 } from "./styled"
 
 /** Processed AudioParamSetting for display on coordinate system */
 type CoordParam = {
