+++ src/components/AudioGraph/elems/styled.tsx
