+++ src/components/AudioGraph/nodes/Gain.tsx
