+++ src/components/AudioGraph/nodes/Oscillator.tsx
