+++ src/components/AudioGraph/styled.tsx
